requirements.txt
alembic==1.8.0
bcrypt==3.2.0
cachelib==0.9.0
cffi==1.15.0
click==8.0.3
cycler==0.11.0
ephem==4.1.3
Flask==2.0.2
Flask-Bcrypt==0.7.1
Flask-Login==0.5.0
Flask-Migrate==3.1.0
Flask-Session==0.4.0
Flask-SQLAlchemy==2.5.1
Flask-WTF==1.0.0
fonttools==4.29.1
greenlet==1.1.2
gunicorn==20.1.0
importlib-metadata==4.12.0
importlib-resources==5.8.0
itsdangerous==2.0.1
Jinja2==3.0.3
kiwisolver==1.3.2
Mako==1.2.1
MarkupSafe==2.0.1
matplotlib==3.5.1
numpy==1.22.1
packaging==21.3
Pillow==9.0.0
psycopg2-binary==2.9.3
pycparser==2.21
pyephem==9.99
pyerfa==2.0.0.1
pyparsing==3.0.7
python-dateutil==2.8.2
python-dotenv==0.20.0
pytz==2021.3
PyYAML==6.0
six==1.16.0
SQLAlchemy==1.4.31
Werkzeug==2.0.2
WTForms==3.0.1
zipp==3.8.0
